Fix refactoring conflicts in Greenplum plugin
- Move Greenplum schema and schema cache to the GP plugin module
